drivers: rtc: maxim,ds3231: implement basic get_time, set_time
Reimplemented the DS1307 driver for the DS3231.
The DS1307's basic functionality is the same as the DS3231 so
it was easy to use that driver as a base.
It was assumed that these were the most critical and easiest
features to implement.
Tests were not yet ran, but a read_time sample adapted from
another PR was added to test partial functionality.

Signed-off-by: Gergo Vari <work@varigergo.hu>